requirements.txt
urllib3==1.26.9
selenium==4.7.2
robotframework==5.0.1
robotframework-requests==0.9.3  # Would like to clean this one in favor of requests
PyYAML==6.0
jsonschema==4.6.1
jsonref-ap==0.3.dev0
pyodbc==4.0.32
geopy==2.2.0
requests==2.28.0
simple-salesforce==1.11.6
phonenumbers==8.12.50
qtdigest==0.3.1
boto3==1.24.19
polyline==1.4.0
lxml==4.9.1
pylint==2.14.4
robotframework-pabot==2.5.3
deepdiff==5.8.1
PyJWT==2.4.0
numpy==1.23.0
openapi-schema-validator==0.3.0
unidecode==1.3.4
pdfminer.six==20220524
openpyxl==3.1.1
pandas==1.5.3
